requirements.txt
boto==2.39.0
coverage==4.0.3
Django==1.9.6
django-coverage==1.2.2
django-debug-toolbar==1.4
django-extensions==1.6.1
django-nose==0.1.3
djangorestframework==3.3.2
django-rest-auth==0.7.0
gunicorn==19.4.5
Jinja2==2.8
MarkupSafe==0.23
mock==0.8.0
MySQL-python==1.2.5
nose==1.3.7
nosexcover==1.0.7
oauthlib==1.0.3
Pillow==3.1.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
sqlparse==0.1.18
newrelic==2.62.0.47
gevent==1.1.1
greenlet==0.4.9
sendgrid==2.0.0
django-post-office==2.0.7